examples/chatserver/tests/requirements.txt
nose
django-nose
mock
mockredispy
websocket-client
